pyproject.toml
[project]
name = "ppo_nnx"
version = "0.1.0"
description = "PPO NNX implementation"
requires-python = ">=3.11, <3.13"
dependencies = [
    "jax[cuda12]>=0.4.34, <0.7.0",
    "numpy>=2.1.2",
    "orbax-checkpoint==0.11.21", # bump when updating brax
    "brax",
    "matplotlib",
    "mediapy",
    "etils",
    "playground",
    "wandb",
    "tyro",
]
[tool.setuptools]
py-modules = []